Move data processing functions to separate file
