+++ components/KenyaContainer/KenyaContainer.tsx
@@ -4,9 +4,16 @@ import { Map } from '../Map/Map';
 import kenyanCounties from './geoJSON/kenyan-counties.json';
 import kenyaSubcounties from './geoJSON/kenya-subcounty-proposal.json';
 import kenyaCountyPopulationData from './geoJSON/county_population_data.json';
+import {
+  findSelectedCountySubcounties,
+  getCenterOfFeatureCollection,
+  getCenterOfSubcountyFeatureCollection,
+  loadCountySelect,
+  loadSubcountySelect,
+  mergeCountiesToPopulation
+} from './KenyaDataManager';
 import { Legend, LegendItem } from '../Legend';
 import * as distance from 'jaro-winkler';
-import * as turf from '@turf/turf';
 import L from 'leaflet';
 
 interface MapContainerProps {
@@ -80,62 +87,6 @@ const KenyaContainer: FunctionComponent<MapContainerProps> = ({ padding }) => {
         counties_with_population: mergeCountiesToPopulation(kenyanCounties, kenyaCountyPopulationData)
       };
     });
-  }
-
-  function mergeCountiesToPopulation(counties: any, countyPopulation: any) {
-    const data: any[] = [];
-    for (const key1 in counties.features) {
-      if (counties.features[key1]) {
-        for (const key2 in countyPopulation) {
-          if (countyPopulation[key2]) {
-            const similarity = distance(
-              counties.features[key1].properties.COUNTY.trim().toLowerCase(),
-              countyPopulation[key2].county.trim().toLowerCase()
-            );
-            if (similarity > 0.7) {
-              const holder = counties.features[key1];
-              data.push({
-                ...holder,
-                properties: {
-                  ...holder.properties,
-                  population: countyPopulation[key2].population
-                }
-              });
-            }
-          }
-        }
-      }
-    }
-
-    return data;
-  }
-
-  function loadCountySelect(counties: any) {
-    const options = [];
-    for (const county in counties.features) {
-      if (counties.features) {
-        options.push({
-          value: counties.features[county].properties.COUNTY,
-          label: counties.features[county].properties.COUNTY
-        });
-      }
-    }
-
-    return options;
-  }
-
-  function loadSubcountySelect(subcounties: any[]) {
-    const options = [];
-    for (const subcounty in subcounties) {
-      if (subcounties[subcounty]) {
-        options.push({
-          value: subcounties[subcounty].properties.ADMIN2,
-          label: subcounties[subcounty].properties.ADMIN2
-        });
-      }
-    }
-
-    return options;
   }
 
   function handleCountyChange(selectedOption: any) {
@@ -230,6 +181,16 @@ const KenyaContainer: FunctionComponent<MapContainerProps> = ({ padding }) => {
     addPopup(latLng, e);
   }
 
+  function addPopup(LatLng: L.LatLng, e: L.LayerEvent) {
+    const popup: L.Popup = L.popup({ autoClose: false })
+    .setLatLng(LatLng)
+    .setContent('<b>County Population</b><br/>' + e.target.feature.properties.COUNTY + '<br/>' +
+    e.target.feature.properties.population);
+    if (state.map) {
+      state.map.addLayer(popup);
+    }
+  }
+
   function resetFeature(e: any) {
       if (state.map) {
         state.map.eachLayer(layer => {
@@ -265,16 +226,6 @@ const KenyaContainer: FunctionComponent<MapContainerProps> = ({ padding }) => {
     }
   }
 
-  function addPopup(LatLng: L.LatLng, e: L.LayerEvent) {
-    const popup: L.Popup = L.popup(/* {autoClose:false} */)
-    .setLatLng(LatLng)
-    .setContent('<b>County Population</b><br/>' + e.target.feature.properties.COUNTY + '<br/>' +
-    e.target.feature.properties.population);
-    if (state.map) {
-      state.map.addLayer(popup);
-    }
-  }
-
   function showKenyaSubcounties() {
     const subCounties = findSelectedCountySubcounties(state.selectedCounty, kenyaSubcounties);
     clean_map();
@@ -299,82 +250,6 @@ const KenyaContainer: FunctionComponent<MapContainerProps> = ({ padding }) => {
         }
       }
     }
-  }
-
-  function findSelectedCountySubcounties(district: string, allSubcounties: any) {
-    const selectedGeometry = [];
-    const subcounties = allSubcounties.features;
-    for (const subcounty in subcounties) {
-      if (subcounties[subcounty]) {
-        const current_district = subcounties[subcounty].properties.ADMIN1;
-        const similarity = distance(district.toLowerCase(), current_district.toLowerCase());
-        if (similarity === 1) {
-          selectedGeometry.push(subcounties[subcounty]);
-        }
-      }
-    }
-
-    return selectedGeometry;
-  }
-
-  function getCenterOfFeatureCollection(subCounties: any[]) {
-    const points: any[] = [];
-    let coordinates_array: any[] = [];
-    for (const key in subCounties) {
-      if (subCounties[key]) {
-        if (subCounties[key].geometry.type === 'Polygon') {
-          coordinates_array = subCounties[key].geometry.coordinates.reduce((p: any, c: any) => {
-            return p.concat(c);
-          });
-        } else if (subCounties[key].geometry.type === 'MultiPolygon') {
-          for (const item in subCounties[key].geometry.coordinates) {
-            if (subCounties[key].geometry.coordinates[item]) {
-              const holder = subCounties[key].geometry.coordinates[item].reduce((p: any, c: any) => {
-                return p.concat(c);
-              });
-              coordinates_array.concat(holder);
-            }
-          }
-
-        }
-
-        for (const item in coordinates_array) {
-          if (coordinates_array[item] instanceof Array) {
-            points.push(turf.point(coordinates_array[item]));
-          }
-        }
-      }
-    }
-
-    return turf.center(turf.featureCollection(points));
-  }
-
-  function getCenterOfSubcountyFeatureCollection(subCounties: any) {
-    const points = [];
-    let coords: any[] = [];
-    if (subCounties.geometry.type === 'Polygon') {
-      coords = subCounties.geometry.coordinates.reduce((p: any, c: any) => {
-        return p.concat(c);
-      });
-    } else if (subCounties.geometry.type === 'MultiPolygon') {
-      for (const item in subCounties.geometry.coordinates) {
-        if (subCounties.geometry.coordinates[item]) {
-          const holder = subCounties.geometry.coordinates[item].reduce((p: any, c: any) => {
-            return p.concat(c);
-          });
-          coords.concat(holder);
-        }
-      }
-
-    }
-    for (const item in coords) {
-      if (coords[item] && (coords[item] instanceof Array)) {
-        points.push(turf.point(coords[item]));
-      }
-    }
-    const features = turf.featureCollection(points);
-
-    return turf.center(features);
   }
 
   return (
